+++ util/annotations/annotations.go
@@ -362,13 +362,14 @@ type HistogramOperation string
 const (
 	HistogramAdd HistogramOperation = "addition"
 	HistogramSub HistogramOperation = "subtraction"
+	HistogramAgg HistogramOperation = "aggregation"
 )
 
 // NewHistogramCounterResetCollisionWarning is used when two counter histograms are added or subtracted where one has
 // a CounterReset hint and the other has NotCounterReset.
 func NewHistogramCounterResetCollisionWarning(pos posrange.PositionRange, operation HistogramOperation) error {
 	switch operation {
-	case HistogramAdd, HistogramSub:
+	case HistogramAdd, HistogramSub, HistogramAgg:
 	default:
 		operation = "unknown operation"
 	}