+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * [BUGFIX] OTLP receiver: Generate `target_info` samples between the earliest and latest samples per resource. #16737
 * [BUGFIX] Config: Infer escaping scheme when scrape config validation scheme is set.
+* [BUGFIX] PromQL: Fix info function on churning series. #17135
 
 ## 3.5.0 / 2025-07-14